toggle subscription on staking_api toggle
