+++ packages/app/src/contexts/Api/index.tsx
@@ -10,7 +10,6 @@ import { Era } from 'api/query/era'
 import { NetworkMeta } from 'api/query/networkMeta'
 import { ActiveEra } from 'api/subscribe/activeEra'
 import { BlockNumber } from 'api/subscribe/blockNumber'
-import { ErasRewardPoints } from 'api/subscribe/erasRewardPoints'
 import { NetworkMetrics } from 'api/subscribe/networkMetrics'
 import { PoolsConfig } from 'api/subscribe/poolsConfig'
 import type { APIEventDetail, ApiStatus, ConnectionType } from 'api/types'
@@ -125,9 +124,12 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
   )
   const stakingMetricsRef = useRef(stakingMetrics)
 
-  // Store chain specs from PAPI
+  // Store chain specs
   const [chainSpecs, setChainSpecs] =
     useState<PapiChainSpecContext>(defaultChainSpecs)
+
+  // Whether the api is ready for querying
+  const isReady = apiStatus === 'ready' && chainSpecs.received === true
 
   // Bootstrap app-wide chain state
   const bootstrapNetworkConfig = async () => {
@@ -399,23 +401,6 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
           activeEraRef
         )
       }
-
-      // Update era reward points subscription
-      // Unsubscribe to staking metrics if it exists.
-      const currentEraRewardPointsSub = Subscriptions.get(
-        network,
-        'erasRewardPoints'
-      )
-      if (currentEraRewardPointsSub) {
-        currentEraRewardPointsSub.unsubscribe()
-        Subscriptions.remove(network, 'erasRewardPoints')
-      }
-      // Subscribe to eras reward points for current era
-      Subscriptions.set(
-        network,
-        'erasRewardPoints',
-        new ErasRewardPoints(network, index)
-      )
     }
   }
 
@@ -543,7 +528,7 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
         setConnectionType,
         rpcEndpoint,
         setRpcEndpoint,
-        isReady: apiStatus === 'ready' && chainSpecs.received === true,
+        isReady,
         consts,
         networkMetrics,
         activeEra,
